Refactored Evaluator and binder
Refactored the binder to produce an array of bound statements. This array is the contents of the block statement produced by lowering the initial bind tree.

The evaluator has now been re-implemented to use this new  array of bound statements with the use of the branch statements.

The display visitor for the binder was also re-implemented. It now uses the console directly and colours the bind tree output.